backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
pydantic[email]
